BCHN v27->v29 security backport (tier 1), the C++20 migration it required,
and the build-system fixes that surfaced along the way. See
doc/upstream-security-backlog.md and doc/release-notes.md.

+++ src/qt/bitcoingui.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2011-2019 The Bitcoin Core developers
-// Copyright (c) 2020-2023 The Bitcoin developers
+// Copyright (c) 2020-2025 The Bitcoin developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -237,14 +237,14 @@ void BitcoinGUI::createActions() {
     overviewAction->setStatusTip(tr("Show general overview of wallet"));
     overviewAction->setToolTip(overviewAction->statusTip());
     overviewAction->setCheckable(true);
-    overviewAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_1));
+    overviewAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_1)));
     tabGroup->addAction(overviewAction);
 
     sendCoinsAction = new QAction(platformStyle->SingleColorIcon(":/icons/send"), tr("&Send"), this);
     sendCoinsAction->setStatusTip(tr("Send coins to a BitFinite address"));
     sendCoinsAction->setToolTip(sendCoinsAction->statusTip());
     sendCoinsAction->setCheckable(true);
-    sendCoinsAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_2));
+    sendCoinsAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_2)));
     tabGroup->addAction(sendCoinsAction);
 
     sendCoinsMenuAction =
@@ -258,7 +258,7 @@ void BitcoinGUI::createActions() {
                                          .arg(QString::fromStdString(config->GetChainParams().CashAddrPrefix())));
     receiveCoinsAction->setToolTip(receiveCoinsAction->statusTip());
     receiveCoinsAction->setCheckable(true);
-    receiveCoinsAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_3));
+    receiveCoinsAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_3)));
     tabGroup->addAction(receiveCoinsAction);
 
     receiveCoinsMenuAction = new QAction(platformStyle->TextColorIcon(":/icons/receiving_addresses"),
@@ -270,7 +270,7 @@ void BitcoinGUI::createActions() {
     historyAction->setStatusTip(tr("Browse transaction history"));
     historyAction->setToolTip(historyAction->statusTip());
     historyAction->setCheckable(true);
-    historyAction->setShortcut(QKeySequence(Qt::ALT + Qt::Key_4));
+    historyAction->setShortcut(QKeySequence(Qt::ALT + static_cast<int>(Qt::Key_4)));
     tabGroup->addAction(historyAction);
 
 #ifdef ENABLE_WALLET
@@ -293,7 +293,7 @@ void BitcoinGUI::createActions() {
 
     quitAction = new QAction(platformStyle->TextColorIcon(":/icons/quit"), tr("E&xit"), this);
     quitAction->setStatusTip(tr("Quit application"));
-    quitAction->setShortcut(QKeySequence(Qt::CTRL + Qt::Key_Q));
+    quitAction->setShortcut(QKeySequence(Qt::CTRL + static_cast<int>(Qt::Key_Q)));
     quitAction->setMenuRole(QAction::QuitRole);
     aboutAction = new QAction(QIcon(":/icons/about"), tr("&About %1").arg(PACKAGE_NAME), this);
     aboutAction->setStatusTip(tr("Show information about %1").arg(PACKAGE_NAME));
@@ -413,12 +413,16 @@ void BitcoinGUI::createMenuBar() {
     QAction *minimize_action = window_menu->addAction(tr("Minimize"));
     minimize_action->setStatusTip(tr("Minimize the Main Window"));
     minimize_action->setToolTip(minimize_action->statusTip());
-    minimize_action->setShortcut(QKeySequence(Qt::CTRL + Qt::Key_M));
-    connect(minimize_action, &QAction::triggered, [] { QApplication::activeWindow()->showMinimized(); });
-    connect(qApp, &QApplication::focusWindowChanged, [minimize_action](QWindow *window) {
-        minimize_action->setEnabled(window != nullptr && (window->flags() & Qt::Dialog) != Qt::Dialog &&
-                                    window->windowState() != Qt::WindowMinimized);
-    });
+    minimize_action->setShortcut(QKeySequence(Qt::CTRL + static_cast<int>(Qt::Key_M)));
+    connect(minimize_action, &QAction::triggered,
+            [] { QApplication::activeWindow()->showMinimized(); });
+    connect(qApp, &QApplication::focusWindowChanged,
+            [minimize_action](QWindow *window) {
+                minimize_action->setEnabled(
+                    window != nullptr &&
+                    (window->flags() & Qt::Dialog) != Qt::Dialog &&
+                    window->windowState() != Qt::WindowMinimized);
+            });
 
 #ifdef Q_OS_MAC
     QAction *zoom_action = window_menu->addAction(tr("Zoom"));
